Documentation deployment using sphinx
Since we write all the docstrings in the numpy format, we can use the numpydoc extension for sphinx to autogenerate the documentation.

For now, it's only the docstrings. We can put more information on usage and other things by including more .rst files as needed.

Lastly, it uses github actions to deploy the static html documentation.